[VideoDatabase] Move cXX column definitions to a separate file
