+++ xbmc/utils/DatabaseUtils.cpp
@@ -14,6 +14,7 @@
 #include "utils/Variant.h"
 #include "utils/log.h"
 #include "video/VideoDatabase.h"
+#include "video/VideoDatabaseColumns.h"
 
 #include <sstream>
